[Feat] #14 - Waiting Card 바뀐 부분 반영
