+++ apps/service/src/components/_refact/waitingCard/WaitingCardLayout.tsx
