+++ apps/service/src/components/_refact/waitingCard/WaitingCardNoCard.tsx
