+++ apps/service/src/components/_refact/waitingCard/_hooks/useWaitingCard.tsx
