+++ apps/service/src/components/_refact/waitingCard/interfaces/waitingCard.ts
