+++ apps/service/src/constants/style.ts
@@ -1,9 +1,9 @@
 export const MAIN_NAVIGATION_HEIGHT = {
-  fold: "3.5rem",
-  unfold: "18.5rem",
+  fold: "3rem",
+  unfold: "13rem",
 };
 
 export const MAIN_FIXED_COMPONENTS_HEIGHT = {
-  fold: "11rem",
-  unfold: "26rem",
+  fold: "8rem",
+  unfold: "18rem",
 };
