+++ apps/service/src/lib/interfaces/waiting.ts
@@ -4,19 +4,19 @@ import { BoothSummary } from "@interfaces/booth";
 export interface WaitingSummary {}
 
 export interface Waiting {
-  waitingID?: number;
-  waitingStatus?: WaitingStatus;
+  waitingID: number;
+  waitingStatus: WaitingStatus;
 
-  waitingCount?: number;
-  totalWaitingTeams?: number;
-  waitingTeamsAhead?: number;
+  waitingCount: number;
+  totalWaitingTeams: number;
+  waitingTeamsAhead: number;
 
-  partySize?: number;
+  partySize: number;
 
-  confirmDueTime?: string;
-  arrivalarrivalDueTime?: string;
+  confirmDueTime: string;
+  arrivalarrivalDueTime: string;
 
-  booth?: BoothSummary;
+  booth: BoothSummary;
 
   // registeredAt: string;
   readyToConfirmAt?: string;
